requirements.txt
ultralytics>=8.3,<9
opencv-python>=4.10,<5
torch>=2.4,<3
torchvision>=0.19,<1
Flask>=3.0,<4
google-genai>=1.0,<2
